Renamed pyVHDLModel.VHDLModel to pyVHDLModel.SyntaxModel. Resolved circular dependency problem.
