+++ doc/Dependency.rst
@@ -3,6 +3,22 @@
 Dependency
 ##########
 
+.. |img-pyVHDLModel-lib-status| image:: https://img.shields.io/librariesio/release/pypi/pyVHDLModel
+   :alt: Libraries.io status for latest release
+   :height: 22
+   :target: https://libraries.io/github/Paebbels/pyVHDLModel
+.. |img-pyVHDLModel-req-status| image:: https://img.shields.io/requires/github/Paebbels/pyVHDLModel
+   :alt: Requires.io
+   :height: 22
+   :target: https://requires.io/github/Paebbels/pyVHDLModel/requirements/?branch=master
+
++------------------------------------------+------------------------------------------+
+| `Libraries.io <https://libraries.io/>`_  | `Requires.io <https://requires.io/>`_    |
++==========================================+==========================================+
+| |img-pyVHDLModel-lib-status|             | |img-pyVHDLModel-req-status|             |
++------------------------------------------+------------------------------------------+
+
+
 .. _dependency-package:
 
 pyVHDLModel Package
