+++ pyVHDLModel/SyntaxModel.py
@@ -41,9 +41,13 @@ This module contains a document language model for VHDL.
 :license: Apache License, Version 2.0
 """
 # load dependencies
-from enum import Enum, unique, IntEnum
-from pathlib  import Path
-from typing import List, Tuple, Union, Dict, Iterator, Optional, Any
+from pathlib              import Path
+from typing               import List, Tuple, Union, Dict, Iterator, Optional, Any
+
+from pydecor.decorators   import export
+
+from pyVHDLModel          import ModelEntity, NamedEntity, MultipleNamedEntity, LabeledEntity, PrimaryUnit, SecondaryUnit, PossibleReference, Direction, EntityClass, Mode
+from pyVHDLModel.PSLModel import VerificationUnit
 
 try:
 	from typing import Protocol
@@ -51,7 +55,6 @@ except ImportError:
 	class Protocol:
 		pass
 
-from pydecor.decorators import export
 
 __all__ = []
 # __api__ = __all__ # FIXME: disabled due to a bug in pydecors export decorator
@@ -89,193 +92,6 @@ Expression = Union[
 	SignalOrSymbol,
 	'Literal',
 ]
-
-
-@export
-@unique
-class Direction(Enum):
-	"""
-	A ``Direction`` is an enumeration and represents a direction in a range
-	(``to`` or ``downto``).
-	"""
-	To =      0
-	DownTo =  1
-
-	def __str__(self):
-		index: int = self.value
-		return ("to", "downto")[index]       # TODO: check performance
-
-
-@export
-@unique
-class Mode(Enum):
-	"""
-	A ``Mode`` is an enumeration. It represents the direction of data exchange
-	(``in``, ``out``, ...) for objects in generic, port or parameter lists.
-
-	In case no *mode* is define, ``Default`` is used, so the *mode* is inferred
-	from context.
-	"""
-	Default = 0
-	In =      1
-	Out =     2
-	InOut =   3
-	Buffer =  4
-	Linkage = 5
-
-	def __str__(self):
-		index: int = self.value
-		return ("", "in", "out", "inout", "buffer", "linkage")[index]       # TODO: check performance
-
-
-@export
-@unique
-class ObjectClass(Enum):
-	"""
-	An ``ObjectClass`` is an enumeration. It represents an object's class (``constant``,
-	``signal``, ...).
-
-	In case no *object class* is define, ``Default`` is used, so the *object class*
-	is inferred from context.
-	"""
-	Default =    0
-	Constant =   1
-	Variable =   2
-	Signal =     3
-	File =       4
-	Type =       5
-	Procedure =  6
-	Function =   7
-
-
-@export
-@unique
-class EntityClass(Enum):
-	"""
-	A ``Class`` is an enumeration. It represents an object's class (``constant``,
-	``signal``, ...).
-
-	In case no *object class* is define, ``Default`` is used, so the *object class*
-	is inferred from context.
-	"""
-	Entity =        0
-	Architecture =  1
-	Configuration = 2
-	Procedure =     3
-	Function =      4
-	Package =       5
-	Type =          6
-	Subtype =       7
-	Constant =      8
-	Signal =        9
-	Variable =      10
-	Component =     11
-	Label =         12
-	Literal =       13
-	Units =         14
-	Group =         15
-	File =          16
-	Property =      17
-	Sequence =      18
-	View =          19
-	Others  =       20
-
-
-@export
-class PossibleReference(IntEnum):
-	Unknown =         0
-	Library =         2**0
-	Entity =          2**1
-	Architecture =    2**2
-	Component =       2**3
-	Package =         2**4
-	Configuration =   2**5
-	Context =         2**6
-	Type =            2**7
-	Subtype =         2**8
-	ScalarType =      2**9
-	ArrayType =       2**10
-	RecordType =      2**11
-	AccessType =      2**12
-	ProtectedType =   2**13
-	FileType =        2**14
-#	Alias =           2**14   # TODO: Is this needed?
-	Attribute =       2**15
-	TypeAttribute =   2**16
-	ValueAttribute =  2**17
-	SignalAttribute = 2**18
-	RangeAttribute =  2**19
-	ViewAttribute =   2**20
-	Constant =        2**16
-	Variable =        2**17
-	Signal =          2**18
-	File =            2**19
-	Object =          2**20   # TODO: Is this needed?
-	EnumLiteral =     2**21
-	Procedure =       2**22
-	Function =        2**23
-	Label =           2**24
-	View =            2**25
-	SimpleNameInExpression = Constant + Variable + Signal + ScalarType + EnumLiteral + Function
-
-@export
-class ModelEntity:
-	"""
-	``ModelEntity`` is the base class for all classes in the VHDL language model,
-	except for mixin classes (see multiple inheritance) and enumerations.
-
-	Each entity in this model has a reference to its parent entity. Therefore
-	a protected variable :attr:`_parent` is available and a readonly property
-	:attr:`Parent`.
-	"""
-	_parent: 'ModelEntity'      #: Reference to a parent entity in the model.
-
-	def __init__(self):
-		pass
-
-	@property
-	def Parent(self) -> 'ModelEntity':
-		"""Returns a reference to the parent entity."""
-		return self._parent
-
-
-@export
-class NamedEntity:
-	"""
-	A ``NamedEntity`` is a mixin class for all VHDL entities that have identifiers.
-
-	A protected variable :attr:`_identifier` is available to derived classes as well as
-	a readonly property :attr:`Identifier` for public access.
-	"""
-	_identifier: str                  #: The name of a model entity.
-
-	def __init__(self, identifier: str):
-		self._identifier = identifier
-
-	@property
-	def Identifier(self) -> str:
-		"""Returns a model entity's identifier (name)."""
-		return self._identifier
-
-
-@export
-class LabeledEntity:
-	"""
-	A ``LabeledEntity`` is a mixin class for all VHDL entities that can have
-	labels.
-
-	A protected variable :attr:`_label` is available to derived classes as well
-	as a readonly property :attr:`Label` for public access.
-	"""
-	_label: str                 #: The label of a model entity.
-
-	def __init__(self, label: str):
-		self._label = label
-
-	@property
-	def Label(self) -> str:
-		"""Returns a model entity's label."""
-		return self._label
 
 
 @export
@@ -697,6 +513,7 @@ class Document(ModelEntity):
 	A ``Document`` represents a sourcefile. It contains primary and secondary
 	design units.
 	"""
+
 	_path:              Path                      #: path to the document. ``None`` if virtual document.
 	_contexts:          List['Context']           #: List of all contexts defined in a document.
 	_configurations:    List['Configuration']     #: List of all configurations defined in a document.
@@ -1202,7 +1019,7 @@ class BitStringLiteral(Literal):
 class ParenthesisExpression(Protocol):
 	@property
 	def Operand(self) -> Expression:
-		...
+		return None
 
 
 @export
@@ -1615,6 +1432,7 @@ class SubtypeAllocation(Allocation):
 	_subtype: Symbol
 
 	def __init__(self, subtype: Symbol):
+		super().__init__()
 		self._subtype = subtype
 
 	@property
@@ -1630,6 +1448,7 @@ class QualifiedExpressionAllocation(Allocation):
 	_qualifiedExpression: QualifiedExpression
 
 	def __init__(self, qualifiedExpression: QualifiedExpression):
+		super().__init__()
 		self._qualifiedExpression = qualifiedExpression
 
 	@property
@@ -1802,12 +1621,12 @@ class RangeSubtype(BaseConstraint):
 
 
 @export
-class Obj(ModelEntity, NamedEntity):
+class Obj(ModelEntity, MultipleNamedEntity):
 	_subtype: SubtypeOrSymbol
 
-	def __init__(self, identifier: str, subtype: SubtypeOrSymbol):
+	def __init__(self, identifiers: List[str], subtype: SubtypeOrSymbol):
 		super().__init__()
-		NamedEntity.__init__(self, identifier)
+		MultipleNamedEntity.__init__(self, identifiers)
 
 		self._subtype = subtype
 
@@ -1839,8 +1658,8 @@ class BaseConstant(Obj):
 
 @export
 class Constant(BaseConstant, WithDefaultExpressionMixin):
-	def __init__(self, identifier: str, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype)
+	def __init__(self, identifiers: List[str], subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype)
 		WithDefaultExpressionMixin.__init__(self, defaultExpression)
 
 
@@ -1848,8 +1667,8 @@ class Constant(BaseConstant, WithDefaultExpressionMixin):
 class DeferredConstant(BaseConstant):
 	_constantReference: Constant
 
-	def __init__(self, identifier: str, subtype: SubtypeOrSymbol):
-		super().__init__(identifier, subtype)
+	def __init__(self, identifiers: List[str], subtype: SubtypeOrSymbol):
+		super().__init__(identifiers, subtype)
 
 	@property
 	def ConstantReference(self) -> Constant:
@@ -1858,8 +1677,8 @@ class DeferredConstant(BaseConstant):
 
 @export
 class Variable(Obj, WithDefaultExpressionMixin):
-	def __init__(self, identifier: str, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype)
+	def __init__(self, identifiers: List[str], subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype)
 		WithDefaultExpressionMixin.__init__(self, defaultExpression)
 
 
@@ -1870,8 +1689,8 @@ class SharedVariable(Obj):
 
 @export
 class Signal(Obj, WithDefaultExpressionMixin):
-	def __init__(self, identifier: str, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super(Signal, self).__init__(identifier, subtype)
+	def __init__(self, identifiers: List[str], subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super(Signal, self).__init__(identifiers, subtype)
 		WithDefaultExpressionMixin.__init__(self, defaultExpression)
 
 
@@ -2066,8 +1885,8 @@ class ParameterInterfaceItem(InterfaceItem):
 
 @export
 class GenericConstantInterfaceItem(Constant, GenericInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifier: str, mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype, defaultExpression)
+	def __init__(self, identifiers: List[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype, defaultExpression)
 		GenericInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
 
@@ -2107,39 +1926,39 @@ class GenericPackageInterfaceItem(GenericInterfaceItem):
 
 @export
 class PortSignalInterfaceItem(Signal, PortInterfaceItem):
-	def __init__(self, identifier: str, mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype, defaultExpression)
+	def __init__(self, identifiers: List[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype, defaultExpression)
 		PortInterfaceItem.__init__(self, mode)
 
 
 @export
 class ParameterConstantInterfaceItem(Constant, ParameterInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifier: str, mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype, defaultExpression)
+	def __init__(self, identifiers: List[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype, defaultExpression)
 		ParameterInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
 
 
 @export
 class ParameterVariableInterfaceItem(Variable, ParameterInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifier: str, mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype, defaultExpression)
+	def __init__(self, identifiers: List[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype, defaultExpression)
 		ParameterInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
 
 
 @export
 class ParameterSignalInterfaceItem(Signal, ParameterInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifier: str, mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
-		super().__init__(identifier, subtype, defaultExpression)
+	def __init__(self, identifiers: List[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: Expression = None):
+		super().__init__(identifiers, subtype, defaultExpression)
 		ParameterInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
 
 
 @export
 class ParameterFileInterfaceItem(File, ParameterInterfaceItem):
-	def __init__(self, identifier: str, subtype: SubtypeOrSymbol):
-		super().__init__(identifier, subtype)
+	def __init__(self, identifiers: List[str], subtype: SubtypeOrSymbol):
+		super().__init__(identifiers, subtype)
 		ParameterInterfaceItem.__init__(self)
 
 # class GenericItem(ModelEntity):
@@ -2219,17 +2038,6 @@ class ContextStatement(Reference):
 
 
 @export
-class DesignUnit(ModelEntity, NamedEntity):
-	"""
-	A ``DesignUnit`` is a base-class for all design units.
-	"""
-
-	def __init__(self, identifier: str):
-		super().__init__()
-		NamedEntity.__init__(self, identifier)
-
-
-@export
 class MixinDesignUnitWithContext:
 	"""
 	A ``DesignUnitWithReferences`` is a base-class for all design units with contexts.
@@ -2254,20 +2062,6 @@ class MixinDesignUnitWithContext:
 	@property
 	def ContextReferences(self) -> List['Context']:
 		return self._contextReferences
-
-
-@export
-class PrimaryUnit(DesignUnit):
-	"""
-	A ``PrimaryUnit`` is a base-class for all primary units.
-	"""
-
-
-@export
-class SecondaryUnit(DesignUnit):
-	"""
-	A ``SecondaryUnit`` is a base-class for all secondary units.
-	"""
 
 
 @export
